Osa4 tehtävä 4.21 tehty

+++ Osa4/blogilista/controllers/blogs.js
@@ -65,8 +65,15 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndDelete(request.params.id)
-    return response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (user.id.toString() === blog.user.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      return response.status(204).end()
+    }
+    return response.status(401).end()
   } catch (error) {
     next(error)
   }
